refactor(unity-ipc): isolate dispatch and failure boundaries
Why:
- Ultra-review found that IPC clients depended back on the execution orchestration namespace and that failure outcome mapping was duplicated.

What Changed:
- Moved dispatch request, IPC request factory, response factory, and failure classifier into dedicated IPC dispatch and failure namespaces.
- Centralized application failure outcome resolution in Application.Shared and reused it from Unity request failures and request result policy.

Impact:
- IPC client dependencies no longer point back at execution orchestration, and plan-token validation failures keep the same application outcome across adapters and request results.
