+++ src/Ucli/UnityIntegration/Ipc/Execution/UnityIpcPluginVerifier.cs
@@ -1,5 +1,6 @@
 using MackySoft.Ucli.Application.Shared.Execution.UnityRequest;
 using MackySoft.Ucli.Application.Shared.Foundation;
+using MackySoft.Ucli.UnityIntegration.Ipc.Failures;
 using MackySoft.Ucli.UnityIntegration.Project.Plugin;
 
 namespace MackySoft.Ucli.UnityIntegration.Ipc.Execution;
