+++ src/Ucli/UnityIntegration/Ipc/Execution/UnityIpcRequestBuilder.cs
@@ -2,6 +2,7 @@ using System.Text.Json;
 using MackySoft.Ucli.Application.Shared.Execution.UnityRequest;
 using MackySoft.Ucli.Contracts;
 using MackySoft.Ucli.Contracts.Ipc;
+using MackySoft.Ucli.UnityIntegration.Ipc.Dispatch;
 
 namespace MackySoft.Ucli.UnityIntegration.Ipc.Execution;
